Extract common type definitions
